requirements.txt
#pytest
pytest>=7.0.0
pytest-cov>=7.0.0
pytest-html>=3.1.1
PyYAML>=6.0
#llmperf
requests>=2.10.0
pandas>=2.3.0
pydantic>=2.12.0
transformers>=4.0.0
httpx>=0.25.0